Convert kts files
